// typings.d.ts
export interface Post {
  _id: string
  _createdAt: string
  publishedAt: string
  title: string
  comments: [Comment]
  author: {
    name: string
    image: string
  }
  description: string
  mainImage: string
  assets: {
    url: string
  }
  slug: {
    current: string
  }
  body: [object]
}

export interface Comment {
  approved: boolean
  comment: string
  email: string
  name: string
  post: {
    _ref: string
    _type: string
  }
  _createdAt: string
  publishedAt: string
  _id: string
  _rev: string
  _type: string
  _updatedAt: string
}